feature/ASE-01: renamed folders for the change logs, force move party…

+++ Kurumi-Engine/Game/UI/Overlays/Core/GlobalMessage.cs
@@ -39,7 +39,7 @@ public sealed class GlobalMessage : IUIOverlay
     internal GlobalMessage(
         AssetRegistry assetRegistry, 
         GlobalMessageDefaults globalMessageDefaults, 
-        int timeLimit, 
+        float timeLimit, 
         string message
     )
     {
